requirements.txt
Flask==2.0.3
gunicorn==20.1.0
gevent==21.12.0
greenlet==1.1.3
Werkzeug==2.0.3
python-dotenv==0.19.2
supabase==2.0.0
httpx==0.25.0